animate welcome part and style the box and text/
